Title: 사칙연산 함수, Time: 8ms, Memory: 0MB, Status: Wrong Answer - Codetree

+++ 250625/사칙연산 함수/quadratic-operations-function.cpp
@@ -8,7 +8,7 @@ char o;
 void solve(int a, int b, char c){
     if(c == '+') cout<<a<<" + "<<b<<" = "<<a+b;
     else if(c == '*') cout<<a<<" * "<<b<<" = "<<a*b;
-    else if(c == '/') cout<<a<<" / "<<b<<" = "<<a/b;
+    else if(c == '/') cout<<a<<" / "<<b<<" = "<<(int)a/b;
     else if(c == '-') cout<<a<<" - "<<b<<" = "<<a-b;
     else cout<<"false";
 }
